signos-agentic/scripts/requirements.txt
PyPDF2>=3.0.0
pdfplumber>=0.10.0
requests>=2.31.0
python-dotenv>=1.0.0
numpy>=1.24.0
tiktoken>=0.5.0

